Moved tests to a separate sub package
